+++ TicTacToe++/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Board"></activity>
+        <activity android:name=".ui.UI" />
     </application>
 
 </manifest>
